Fixed a bug in nineoldandroids that caused the app to crash upon animating ints. Performance seems fine on the S2 with 2.3

+++ library/src/com/nineoldandroids/animation/PropertyValuesHolder.java
@@ -828,7 +828,7 @@ public class PropertyValuesHolder implements Cloneable {
                 return;
             }
             if (mProperty != null) {
-                mProperty.set(target, mIntAnimatedValue);
+                mProperty.set(target, (float)mIntAnimatedValue);
                 return;
             }
             //if (mJniSetter != 0) {
